quangnd

+++ app/Jobs/OtpEndTimeJob.php
@@ -3,6 +3,7 @@
 namespace App\Jobs;
 
 use App\Models\OtpVerify;
+use Carbon\Carbon;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
@@ -14,6 +15,12 @@ class OtpEndTimeJob implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
     public $otp;
+        /**
+     * Delete the job if its models no longer exist.
+     *
+     * @var bool
+     */
+    public $deleteWhenMissingModels = true;
     /**
      * Create a new job instance.
      */
@@ -27,14 +34,18 @@ class OtpEndTimeJob implements ShouldQueue
      */
     public function handle(): void
     {
-      
-        if($this->otp){
-            $otp = OtpVerify::find($this->otp->id);
-         
-            if($otp){
-                $otp->delete();
+        if($this->otp->expried_at == Carbon::now()){
+            if($this->otp){
+                $otp = OtpVerify::find($this->otp->id);
+             
+                if($otp){
+                    $otp->delete();
+                }
             }
         }
+     
        
     }
+
+
 }
